Split out `temporal-versioning` from `hash-graph` into crate
