+++ apps/hash-graph/Cargo.toml
@@ -14,9 +14,20 @@ publish = false
 repository = "https://github.com/hashintel/hash/tree/main/apps/hash-graph"
 
 [workspace.dependencies]
+# Ideally we would move this to `libs/` but we currently cannot have multiple Rust workspaces depending on each other
+# as otherwise the CI would not be able to determine the dependency order. This would result in dependants not being
+# tested in CI if the dependency was changed.
+temporal-versioning = { path = "lib/temporal-versioning" }
+
 type-system = { git = "https://github.com/blockprotocol/blockprotocol", rev = "cdde49" }
 hash-status = { path = "../../libs/@local/status/crate" }
 error-stack = { version = "0.3.1", features = ["spantrace"] }
+
+utoipa = "3.3.0"
+serde = "1.0.183"
+time = { version = "0.3.23", default-features = false }
+tracing = "0.1.37"
+postgres-types = { version = "0.2.5", default-features = false }
 
 [profile.production]
 inherits = "release"
