+++ apps/hash-graph/lib/graph/Cargo.toml
@@ -6,11 +6,17 @@ publish = false
 description = "HASH Graph API"
 
 [dependencies]
+temporal-versioning = { workspace = true, features = ["postgres", "utoipa"] }
 type-fetcher = { path = "../type-fetcher" }
 
 error-stack = { workspace = true }
 hash-status = { workspace = true }
 type-system = { workspace = true }
+
+postgres-types = { workspace = true, features = ["derive", "with-uuid-1", "with-serde_json-1"] }
+serde = { workspace = true, features = ["derive"] }
+utoipa = { workspace = true, features = ["uuid"] }
+tracing = { workspace = true }
 
 async-trait = "0.1.73"
 axum = "0.6.19"
@@ -24,15 +30,12 @@ include_dir = "0.7.3"
 mime = "0.3.17"
 opentelemetry = { version = "0.18.0", features = ["rt-tokio"] }
 opentelemetry-otlp = "0.11.0"
-postgres-protocol = "0.6.5"
-postgres-types = { version = "0.2.5", default-features = false, features = ["derive", "with-uuid-1", "with-serde_json-1", "with-time-0_3"] }
 refinery = { version = "0.8", features = ["tokio-postgres"] }
 regex = "1.9.1"
 semver = { version = "1.0.18", default-features = false, features = ["serde"] }
-serde = { version = "1.0.183", features = ["derive"] }
 serde_json = "1.0.104"
 tarpc = { version = "0.33", features = ["serde-transport", "tcp"] }
-time = { version = "0.3.23", features = ["serde", "formatting", "macros"] }
+time = { workspace = true }
 tokio = { version = "1.29.1", default-features = false }
 tokio-postgres = { version = "0.7.8", default-features = false }
 tokio-serde = { version = "0.8", features = ["json"] }
@@ -40,12 +43,10 @@ tokio-util = { version = "0.7.8", default-features = false, features = ["codec",
 tonic = "0.8.3"
 tower = "0.4.13"
 tower-http = { version = "0.4.2", features = ["trace"] }
-tracing = "0.1.37"
 tracing-appender = "0.2.2"
 tracing-error = "0.2.0"
 tracing-opentelemetry = "0.18.0"
 tracing-subscriber = { version = "0.3.17", features = ["env-filter", "json"] }
-utoipa = { version = "3.3.0", features = ["uuid"] }
 uuid = { version = "1.4.1", features = ["v4", "serde"] }
 
 [dev-dependencies]
