+++ apps/hash-graph/lib/graph/src/api/rest/utoipa_typedef/subgraph/edges.rs
@@ -4,6 +4,7 @@ use std::{
 };
 
 use serde::Serialize;
+use temporal_versioning::Timestamp;
 use type_system::url::BaseUrl;
 use utoipa::{
     openapi::{schema::AdditionalProperties, ObjectBuilder, OneOfBuilder, Ref, RefOr, Schema},
@@ -12,7 +13,7 @@ use utoipa::{
 
 use crate::{
     api::rest::utoipa_typedef::subgraph::vertices::OntologyTypeVertexId,
-    identifier::{knowledge::EntityId, ontology::OntologyTypeVersion, time::Timestamp},
+    identifier::{knowledge::EntityId, ontology::OntologyTypeVersion},
     subgraph::{
         edges::{KnowledgeGraphEdgeKind, OntologyEdgeKind, OutwardEdge, SharedEdgeKind},
         identifier::{
@@ -240,17 +241,15 @@ impl ToSchema<'_> for Edges {
 
 #[cfg(test)]
 mod tests {
+    use temporal_versioning::{
+        ClosedTemporalBound, LeftClosedTemporalInterval, OpenTemporalBound, Timestamp,
+    };
     use type_system::url::BaseUrl;
     use uuid::Uuid;
 
     use crate::{
         api::rest::utoipa_typedef::subgraph::Edges,
-        identifier::{
-            account::AccountId,
-            knowledge::EntityId,
-            ontology::OntologyTypeVersion,
-            time::{ClosedTemporalBound, LeftClosedTemporalInterval, OpenTemporalBound, Timestamp},
-        },
+        identifier::{account::AccountId, knowledge::EntityId, ontology::OntologyTypeVersion},
         knowledge::EntityUuid,
         provenance::OwnedById,
         subgraph::{
