+++ apps/hash-graph/lib/graph/src/lib.rs
@@ -21,8 +21,6 @@ pub mod knowledge;
 pub mod ontology;
 pub mod subgraph;
 
-mod serde;
-
 pub mod store;
 
 pub mod snapshot;
