+++ apps/hash-graph/lib/graph/src/shared.rs
@@ -1,3 +1,2 @@
 pub mod identifier;
-pub mod interval;
 pub mod provenance;
