+++ apps/hash-graph/lib/graph/src/shared/identifier.rs
@@ -1,4 +1,3 @@
 pub mod account;
 pub mod knowledge;
 pub mod ontology;
-pub mod time;
