+++ apps/hash-graph/lib/graph/src/snapshot/ontology/table.rs
@@ -1,13 +1,11 @@
 use postgres_types::{Json, ToSql};
+use temporal_versioning::{LeftClosedTemporalInterval, TransactionTime};
 use time::OffsetDateTime;
 use type_system::repr;
 use uuid::Uuid;
 
 use crate::{
-    identifier::{
-        ontology::OntologyTypeVersion,
-        time::{LeftClosedTemporalInterval, TransactionTime},
-    },
+    identifier::ontology::OntologyTypeVersion,
     provenance::{OwnedById, RecordArchivedById, RecordCreatedById},
 };
