+++ apps/hash-graph/lib/graph/src/store/knowledge.rs
@@ -1,12 +1,10 @@
 use async_trait::async_trait;
 use error_stack::Result;
+use temporal_versioning::{DecisionTime, Timestamp};
 use type_system::url::VersionedUrl;
 
 use crate::{
-    identifier::{
-        knowledge::EntityId,
-        time::{DecisionTime, Timestamp},
-    },
+    identifier::knowledge::EntityId,
     knowledge::{Entity, EntityLinkOrder, EntityMetadata, EntityProperties, EntityUuid, LinkData},
     provenance::{OwnedById, RecordCreatedById},
     store::{crud, InsertionError, QueryError, UpdateError},
