+++ apps/hash-graph/lib/graph/src/store/postgres/knowledge/entity/read.rs
@@ -3,6 +3,9 @@ use std::{borrow::Cow, mem::swap, str::FromStr};
 use async_trait::async_trait;
 use error_stack::{IntoReport, Result, ResultExt};
 use futures::{StreamExt, TryStreamExt};
+use temporal_versioning::{
+    LeftClosedTemporalInterval, RightBoundedTemporalInterval, TemporalTagged, TimeAxis, Timestamp,
+};
 use tokio_postgres::GenericClient;
 use type_system::url::{BaseUrl, VersionedUrl};
 use uuid::Uuid;
@@ -11,10 +14,6 @@ use crate::{
     identifier::{
         account::AccountId,
         knowledge::{EntityEditionId, EntityId, EntityRecordId, EntityTemporalMetadata},
-        time::{
-            LeftClosedTemporalInterval, RightBoundedTemporalInterval, TemporalTagged, TimeAxis,
-            Timestamp,
-        },
     },
     knowledge::{Entity, EntityLinkOrder, EntityMetadata, EntityQueryPath, EntityUuid, LinkData},
     ontology::EntityTypeQueryPath,
