+++ apps/hash-graph/lib/graph/src/store/postgres/ontology/entity_type.rs
@@ -5,6 +5,7 @@ use async_trait::async_trait;
 use error_stack::IntoReport;
 use error_stack::{Report, Result, ResultExt};
 use futures::{stream, TryStreamExt};
+use temporal_versioning::RightBoundedTemporalInterval;
 use type_system::{
     url::{BaseUrl, VersionedUrl},
     EntityType,
@@ -13,7 +14,7 @@ use type_system::{
 #[cfg(hash_graph_test_environment)]
 use crate::store::error::DeletionError;
 use crate::{
-    identifier::{ontology::OntologyTypeRecordId, time::RightBoundedTemporalInterval},
+    identifier::ontology::OntologyTypeRecordId,
     ontology::{
         EntityTypeMetadata, EntityTypeWithMetadata, OntologyTemporalMetadata,
         PartialCustomEntityTypeMetadata, PartialCustomOntologyMetadata, PartialEntityTypeMetadata,
