+++ apps/hash-graph/lib/graph/src/subgraph/identifier/edge.rs
@@ -1,10 +1,8 @@
 use serde::{Deserialize, Serialize};
+use temporal_versioning::LeftClosedTemporalInterval;
 use utoipa::ToSchema;
 
-use crate::{
-    identifier::{knowledge::EntityId, time::LeftClosedTemporalInterval},
-    subgraph::temporal_axes::VariableAxis,
-};
+use crate::{identifier::knowledge::EntityId, subgraph::temporal_axes::VariableAxis};
 
 pub trait EdgeEndpoint {
     type BaseId;
