+++ apps/hash-graph/lib/temporal-versioning/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "temporal-versioning"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+serde = { workspace = true, features = ["derive"] }
+utoipa = { workspace = true, optional = true }
+postgres-types = { workspace = true, features = ["with-time-0_3"], optional = true }
+time = { workspace = true, features = ["serde", "parsing", "formatting", "macros"] }
+tracing = { workspace = true, optional = true }
+
+derivative = "2.2.0"
+postgres-protocol = { version = "0.6.5", default-features = false, optional = true }
+
+[features]
+postgres = ["dep:postgres-types", "dep:postgres-protocol", "dep:tracing"]
+utoipa = ["dep:utoipa"]
