+++ apps/hash-graph/lib/temporal-versioning/src/serde.rs
