+++ apps/hash-graph/lib/temporal-versioning/src/utoipa.rs
@@ -1,0 +1,3 @@
+use utoipa::{openapi, ToSchema};
+
+use crate::{Interval, Timestamp};
