+++ apps/hash-graph/tests/integration/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 [dev-dependencies]
 graph = { path = "../../lib/graph" }
 graph-test-data = { path = "../test_data" }
+temporal-versioning = { workspace = true }
 
 error-stack = { workspace = true }
 type-system = { workspace = true }
@@ -15,7 +16,7 @@ futures = "0.3.28"
 rand = "0.8.5"
 serde = { version = "1.0.183", features = ["derive"] }
 serde_json = "1.0.104"
-time = "0.3.23"
+time = { workspace = true }
 tokio = { version = "1.29.1", features = ["rt-multi-thread", "macros"] }
 tokio-postgres = { version = "0.7.8", default-features = false }
 uuid = { version = "1.4.1", features = ["v4", "serde"] }
